Added new api to search user | working proto type

+++ src/main/java/de/ij3rry/chatApp/services/impl/ChatServiceImpl.java
@@ -1,5 +1,6 @@
 package de.ij3rry.chatApp.services.impl;
 
+import de.ij3rry.chatApp.authentication.MessageRequestValidator;
 import de.ij3rry.chatApp.components.SendMessageComponent;
 import de.ij3rry.chatApp.converters.DTOToDocumentConverter;
 import de.ij3rry.chatApp.documents.AppUserDocument;
@@ -12,12 +13,19 @@ import de.ij3rry.chatApp.repositories.AppUserRepository;
 import de.ij3rry.chatApp.repositories.InboundMessageRepository;
 import de.ij3rry.chatApp.repositories.OnlineUserRepository;
 import de.ij3rry.chatApp.services.ChatService;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.messaging.MessageHeaders;
+import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
 import java.util.UUID;
 
+@Slf4j
 @Service
 public final class ChatServiceImpl implements ChatService {
 
@@ -37,22 +45,24 @@ public final class ChatServiceImpl implements ChatService {
         this.appUserRepository = appUserRepository;
     }
     @Override
-    public void checkInUser(CheckInDTO checkInDTO) {
+    public void checkInUser(CheckInDTO checkInDTO, MessageHeaders headers) {
+        final UUID privateUserId = MessageRequestValidator.getPrivateUserIDFromSIMPHeader(headers);
 
-        // TODO : For now using a static . need to implement authentication system
-        final UUID privateUserId = UUID.fromString("");
         /* fetch unsent messages from DB */
         final AppUserDocument appUserDocument = appUserRepository.findByPrivateID(privateUserId);
+        if( appUserDocument == null )
+            throw new UsernameNotFoundException("App User not fund!");
+
         final List<InboundMessageDocument> inboundMessageDocuments = inboundMessageRepository.findByToUserID(appUserDocument.getPublicID());
         /* send the message to the online user */
-        if( inboundMessageDocuments !=null ){
+        if( inboundMessageDocuments !=null && !inboundMessageDocuments.isEmpty()){
             inboundMessageDocuments.forEach(document -> {
                 final OutgoingMessageDTO outgoingMessageDTO = new OutgoingMessageDTO(document.getFromUserID(),document.getBody(),document.getType());
                 sendMessageComponent.sendMessagesToUser(checkInDTO.getTopicId(),outgoingMessageDTO);
             });
+            /* then delete the messages which is delivered */
+            inboundMessageRepository.deleteAll(inboundMessageDocuments);
         }
-        /* then delete the messages which is delivered */
-        inboundMessageRepository.deleteAll(inboundMessageDocuments);
 
         /* updating topic id for realtime transaction */
         appUserDocument.setTopicID(checkInDTO.getTopicId());
@@ -61,22 +71,24 @@ public final class ChatServiceImpl implements ChatService {
     }
 
     @Override
-    public void handleIncomingMessage(IncomingMessageDTO incommingMessageDTO) {
-        final AppUserDocument fromUser = appUserRepository.findByPrivateID(incommingMessageDTO.getPrivateUserID());
+    public void handleIncomingMessage(IncomingMessageDTO incomingMessageDTO, MessageHeaders headers) {
+        final UUID privateUserId = MessageRequestValidator.getPrivateUserIDFromSIMPHeader(headers);
+        final AppUserDocument fromUser = appUserRepository.findByPrivateID(privateUserId);
         if( fromUser == null )
-            return;
+            throw new UsernameNotFoundException("From User not fund!");
 
-        final OnlineUserDocument onlineUserDocument = onlineUserRepository.findByPublicUserId(incommingMessageDTO.getToUserID());
         /* check the user is available */
-        if( onlineUserDocument != null && onlineUserDocument.isOnline() ){
-            final OutgoingMessageDTO outgoingMessageDTO = new OutgoingMessageDTO(fromUser.getPublicID(),incommingMessageDTO.getBody(),incommingMessageDTO.getType());
-            /* fetching the updated topicID */
-            final AppUserDocument toUser = appUserRepository.findByPublicID(incommingMessageDTO.getToUserID());
+        final AppUserDocument toUser = appUserRepository.findByPublicID(incomingMessageDTO.getToUserId());
+        if( toUser == null)
+            throw new UsernameNotFoundException("To user not found!");
+        if( toUser.isOnline() ){
+            final OutgoingMessageDTO outgoingMessageDTO = new OutgoingMessageDTO(fromUser.getPublicID(),incomingMessageDTO.getBody(),incomingMessageDTO.getType());
+
             sendMessageComponent.sendMessagesToUser(toUser.getTopicID(),outgoingMessageDTO);
         }
         /* Store the messages for now */
         else{
-            final InboundMessageDocument inboundMessageDocument = DTOToDocumentConverter.covertIncomingMessageDTOToDocument(incommingMessageDTO);
+            final InboundMessageDocument inboundMessageDocument = DTOToDocumentConverter.covertIncomingMessageDTOToDocument(incomingMessageDTO);
             inboundMessageDocument.setFromUserID(fromUser.getPublicID());
             inboundMessageRepository.save(inboundMessageDocument);
         }
